apps/gen_imgs_with_yaml/requirements.txt
litellm
PyYAML
python-dotenv
mcp
